pyproject.toml
[project]
name = "neutron"
version = "0.1.0"
description = ""
authors = [
    {name = "Ikigami0513",email = "ikigami0513@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "jinja2 (>=3.1.6,<4.0.0)",
    "pywebview (>=6.1,<7.0)",
    "flask (>=3.1.2,<4.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
